+++ models_pack/db_access.py
@@ -1,7 +1,7 @@
 #! /usr/bin/env python
 # -*- coding: utf8 -*-
 from models_pack.models import *
-import my_loging
+from models_pack import my_loging
 from datetime import datetime
