+++ uob.py
@@ -1,8 +1,7 @@
 #! /usr/bin/env python
 # -*- coding: utf8 -*-
 import click
-import my_loging
-from models_pack import db_access
+from models_pack import db_access, my_loging
 import os
